+++ CHANGELOG.md
@@ -61,6 +61,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix issue in asset navigators where folders could incorrectly show files as being within folders if filename matched the folder name
 - Fix issue in song navigator where keyboard navigation could stop working
 - Fix BPM label calculation in Music Editor
+- Fix issue where Select All could stop working in Tracker view
 
 ## [4.2.2] - 2026-02-24